refact: 파일구조 index.tsx, style.tsx로 변경
