--- frontend/src/components/common/my-car-image-box/index.tsx
@@ -1,1 +1,0 @@
-export { MyCarImageBox } from './MyCarImageBox';
