--- frontend/src/components/common/trim-card/index.tsx
@@ -1,1 +1,0 @@
-export { TrimCard } from './TrimCard';
